fixed another issue with process_links, added shitty cmd switches.  Added new score fields to es search

+++ scripts/process_links.py
@@ -11,7 +11,7 @@ import re
 
 p = etree.XMLParser(huge_tree=True)
 
-def run(db, config, do_id_lookup=True, do_references=True):
+def run(db, config, do_id_lookup=True, do_references=True, do_subordinates=True):
     if do_id_lookup:
         with db.cursor(cursor_factory=extras.RealDictCursor) as cur:
             cur.execute(""" delete from id_lookup""")
@@ -44,11 +44,12 @@ def run(db, config, do_id_lookup=True, do_references=True):
         if do_references:
             cur.execute(""" delete from document_references""")
             cur.execute(""" delete from section_references""")
-        cur.execute(""" delete from subordinates""")
+        if do_subordinates:
+            cur.execute(""" delete from subordinates""")
 
     with db.cursor(cursor_factory=extras.RealDictCursor, name="law_cursor") as cur:
         result_to_dict = lambda r: (r['govt_id'], r['parent_id'])
-        cur.execute("""SELECT govt_id, parent_id from id_lookup """)
+        cur.execute("""SELECT i.govt_id, parent_id from id_lookup i join latest_instruments l on i.parent_id = l.id """)
         id_lookup = dict(map(result_to_dict, cur.fetchall()))
 
     with db.cursor(cursor_factory=extras.RealDictCursor, name="law_cursor") as cur:
@@ -56,14 +57,14 @@ def run(db, config, do_id_lookup=True, do_references=True):
         cur.execute(query)
 
         titles = {unicode(x['title'].decode('utf-8')): x['id'] for x in cur.fetchall()}
-        id_to_title = {v:k for k, v in titles.items()}
+        id_to_title = {v: k for k, v in titles.items()}
 
         keys = sorted(titles.keys(), key=lambda x: len(x), reverse=True)
         regex = re.compile(u"(^|\W)(%s)($|\W)" % u"|".join(map(lambda x: re.escape(x), keys)), flags=re.I & re.UNICODE)
 
     with db.cursor(cursor_factory=extras.RealDictCursor, name="law_cursor") as cur, db.cursor() as out:
         count = 0
-        cur.execute("""SELECT document, d.id, title from instruments i join documents d on i.id = d.id  """)
+        cur.execute("""SELECT document, d.id, title from instruments i join documents d on i.id = d.id""")
         documents = cur.fetchmany(1)
         while len(documents):
             for document in documents:
@@ -77,7 +78,7 @@ def run(db, config, do_id_lookup=True, do_references=True):
                 if do_references:
                     links = map(lambda x: {'id': x.attrib['href'],
                         'path': generate_path_string(x, title=unicode(document['title'].decode('utf-8')))},
-                        tree.xpath('.//extref[@href]|.//intref[@href]'))
+                        tree.xpath('.//extref[@href][not(ancestor::history-note)]|.//intref[@href][not(ancestor::history-note)]'))
                     counters = defaultdict(int)
                     for link in links:
                         if link['id'] in id_lookup:
@@ -90,77 +91,77 @@ def run(db, config, do_id_lookup=True, do_references=True):
                         flattened = map(lambda x: (source_id, x['id'], x['path'][0], x['path'][1]), links)
                         args_str = ','.join(cur.mogrify("(%s,%s,%s,%s)", x) for x in flattened)
                         out.execute("INSERT INTO section_references (source_document_id, target_govt_id, repr, url) VALUES " + args_str)
-
-                ids = []
-                pursuant = tree.xpath('.//pursuant[not(ancestor::end)][not(ancestor::skeleton)]')
-                title = unicode(document['title'].decode('utf-8'))
-                if len(pursuant):
-                    try:
-                        refs = pursuant[0].xpath('.//extref')
-                        if len(refs):
-                            ids = [id_lookup[ref.attrib['href']] for ref in refs]
-                        else:
-                            ids = [titles[x[1]] for x in regex.findall(etree.tostring(pursuant[0], method="text", encoding="UTF-8"))]
-                        ids = list(set(ids))
-                    except KeyError:
-                        print refs, document.get('title')
-                else:
-                    # see if there is a 'is called the principal Act'
-                    rules = ('translate(., "ABCDEFGHIJKLMNOPQRSTUVWXYZ", "abcdefghijklmnopqrstuvwxyz")', '[not(ancestor::end)][not(ancestor::skeleton)]')
-                    principal = tree.xpath('//text[contains(%s, "is called the principal act")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "these regulations amend the")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "an act to amend the")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "this act amends the")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "these regulations amend the")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "this part amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "these parts amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "this section amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "these sections amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "this clause amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "this clauses amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "this schedule amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "these schedules amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "this regulation amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "this rule amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//text[contains(%s, "these rules amends ")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//def-term[contains(%s, "principal act")]%s' % rules)
-                    if not len(principal):
-                        principal = tree.xpath('//def-term[contains(%s, "principal regulations")]%s' % rules)
-                    try:
-                        if len(principal):
-                            link = principal[0].xpath('./extref|./def-term')
-                            if len(link):
-                                link_id = link[0].attrib.get('href') or link[0].attrib.get('id')
-                                ids = [id_lookup[link_id]]
+                if do_subordinates:
+                    ids = []
+                    pursuant = tree.xpath('.//pursuant[not(ancestor::end)][not(ancestor::skeleton)]')
+                    title = unicode(document['title'].decode('utf-8'))
+                    if len(pursuant):
+                        try:
+                            refs = pursuant[0].xpath('.//extref')
+                            if len(refs):
+                                ids = [id_lookup[ref.attrib['href']] for ref in refs]
                             else:
-                                ids = [titles[x[1]] for x in regex.findall(etree.tostring(principal[0], method="text", encoding="UTF-8"))]
-                    except (KeyError, IndexError):
-                        print document.get('title')
-                    # 'These regulations amend the'
-                ids = list(set([i for i in ids if i != document.get('id') and id_to_title.get(i) != title]))
-                if len(ids):
-                    print document.get('title'), [id_to_title.get(i) for i in ids]
-                    args_str = ','.join(cur.mogrify("(%s, %s)", (x, document['id'])) for x in ids)
-                    out.execute("INSERT INTO subordinates (parent_id, child_id) VALUES " + args_str)
-                else:
-                    if 'Amendment' in document.get('title'):
-                        print document.get('title')
+                                ids = [titles[x[1]] for x in regex.findall(etree.tostring(pursuant[0], method="text", encoding="UTF-8"))]
+                            ids = list(set(ids))
+                        except KeyError:
+                            print refs, document.get('title')
+                    else:
+                        # see if there is a 'is called the principal Act'
+                        rules = ('translate(., "ABCDEFGHIJKLMNOPQRSTUVWXYZ", "abcdefghijklmnopqrstuvwxyz")', '[not(ancestor::end)][not(ancestor::skeleton)]')
+                        principal = tree.xpath('//text[contains(%s, "is called the principal act")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "these regulations amend the")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "an act to amend the")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "this act amends the")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "these regulations amend the")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "this part amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "these parts amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "this section amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "these sections amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "this clause amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "this clauses amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "this schedule amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "these schedules amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "this regulation amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "this rule amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//text[contains(%s, "these rules amends ")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//def-term[contains(%s, "principal act")]%s' % rules)
+                        if not len(principal):
+                            principal = tree.xpath('//def-term[contains(%s, "principal regulations")]%s' % rules)
+                        try:
+                            if len(principal):
+                                link = principal[0].xpath('./extref|./def-term')
+                                if len(link):
+                                    link_id = link[0].attrib.get('href') or link[0].attrib.get('id')
+                                    ids = [id_lookup[link_id]]
+                                else:
+                                    ids = [titles[x[1]] for x in regex.findall(etree.tostring(principal[0], method="text", encoding="UTF-8"))]
+                        except (KeyError, IndexError):
+                            print document.get('title')
+                        # 'These regulations amend the'
+                    ids = list(set([i for i in ids if i != document.get('id') and id_to_title.get(i) != title]))
+                    if len(ids):
+                        print document.get('title'), [id_to_title.get(i) for i in ids]
+                        args_str = ','.join(cur.mogrify("(%s, %s)", (x, document['id'])) for x in ids)
+                        out.execute("INSERT INTO subordinates (parent_id, child_id) VALUES " + args_str)
+                    else:
+                        if 'Amendment' in document.get('title'):
+                            print 'Could not find parent for: ', document.get('title')
             documents = cur.fetchmany(1)
     db.commit()
     with db.cursor(cursor_factory=extras.RealDictCursor) as cur:
@@ -199,6 +200,7 @@ def run(db, config, do_id_lookup=True, do_references=True):
             cur.execute(query)
             results = cur.fetchall()
             if len(results):
+
                 print 'removing cycle', dict(results[0])
                 cur.execute(rm_query, results[0])
             else:
@@ -220,6 +222,10 @@ if __name__ == "__main__":
             user=config.DB_USER,
             host=config.DB_HOST,
             password=config.DB_PW)
-    run(db, config)
+    # poor mans switches
+    run(db, config,
+        do_id_lookup=('skip_ids' not in sys.argv[1:]),
+        do_references=('skip_references' not in sys.argv[1:]),
+        do_subordinates=('skip_subordinates' not in sys.argv[1:]))
     db.close()
